# packages/api-client/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Smartnet.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/quotations": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationSummary"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationSummary"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuotationSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Quotations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateQuotationRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateQuotationRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateQuotationRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationCreatedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationCreatedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationCreatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotations/{id}": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationDetail"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotationDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotations/tax-rate": {
      "get": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTaxRate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTaxRate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTaxRate"
                }
              }
            }
          }
        }
      }
    },
    "/api/quotations/{id}/convert": {
      "post": {
        "tags": [
          "Quotations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ConvertQuotationRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ConvertQuotationRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ConvertQuotationRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceCreatedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceCreatedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceCreatedResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConvertQuotationRequest": {
        "required": [
          "date",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "purchaseOrderNo": {
            "type": "string",
            "nullable": true
          },
          "contactPerson": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateInvoiceLineRequest": {
        "required": [
          "discountPercent",
          "quantity",
          "unitPrice"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "itemCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "discountPercent": {
            "type": "number",
            "format": "double"
          },
          "cost": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateQuotationRequest": {
        "required": [
          "companyId",
          "customerId",
          "date",
          "documentDiscountPercent",
          "lines"
        ],
        "type": "object",
        "properties": {
          "companyId": {
            "type": "integer",
            "format": "int64"
          },
          "customerId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "contactPerson": {
            "type": "string",
            "nullable": true
          },
          "validity": {
            "type": "string",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInvoiceLineRequest"
            }
          },
          "documentDiscountPercent": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceCreatedResponse": {
        "required": [
          "id",
          "number",
          "outstanding",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "outstanding": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceLineDetail": {
        "required": [
          "discountPercent",
          "gross",
          "net",
          "quantity",
          "unitPrice"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "itemCode": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "discountPercent": {
            "type": "number",
            "format": "double"
          },
          "gross": {
            "type": "number",
            "format": "double"
          },
          "net": {
            "type": "number",
            "format": "double"
          },
          "cost": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceTaxRate": {
        "required": [
          "name",
          "percentage"
        ],
        "type": "object",
        "properties": {
          "taxRateId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "percentage": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "QuotationCreatedResponse": {
        "required": [
          "id",
          "number",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "QuotationDetail": {
        "required": [
          "date",
          "discountAmount",
          "id",
          "kind",
          "lines",
          "netTotal",
          "number",
          "origin",
          "subtotal",
          "taxAmount",
          "taxRatePercentage",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "kind": {
            "type": "string"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerCode": {
            "type": "string",
            "nullable": true
          },
          "contactPerson": {
            "type": "string",
            "nullable": true
          },
          "validity": {
            "type": "string",
            "nullable": true
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "discountAmount": {
            "type": "number",
            "format": "double"
          },
          "netTotal": {
            "type": "number",
            "format": "double"
          },
          "taxRatePercentage": {
            "type": "number",
            "format": "double"
          },
          "taxAmount": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "convertedInvoiceId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "convertedInvoiceNumber": {
            "type": "string",
            "nullable": true
          },
          "origin": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineDetail"
            }
          }
        },
        "additionalProperties": false
      },
      "QuotationSummary": {
        "required": [
          "date",
          "id",
          "number",
          "origin",
          "total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "convertedInvoiceId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "origin": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}